+++ packages/site/src/data/react/libraries.ts
@@ -48,9 +48,31 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		npmPackage: "redux",
 		href: "https://redux.js.org/",
 		description:
-			"A predictable, centralized, debuggable and flexible state container for JS apps inspired by the Flux and Elm architectures.",
+			"A predictable, centralized, debuggable and flexible state container for JS apps inspired by the Flux and Elm architectures. Nowadays you should be using Redux Toolkit instead.",
 		image: "https://github.com/reduxjs.png",
 		tags: ["state management", "redux"],
+	},
+	{
+		name: "Redux Toolkit",
+		author: "Mark Erikson",
+		gitHubRepo: "reduxjs/redux-toolkit",
+		npmPackage: "@reduxjs/toolkit",
+		href: "https://redux-toolkit.js.org/",
+		description:
+			"The official, opinionated, batteries-included toolset for efficient Redux development",
+		image: "https://github.com/reduxjs.png",
+		tags: ["state management", "redux"],
+	},
+	{
+		name: "RTK Query",
+		author: "Lenz Weber",
+		gitHubRepo: "reduxjs/redux-toolkit",
+		npmPackage: "@reduxjs/toolkit",
+		href: "https://redux-toolkit.js.org/rtk-query/overview",
+		description:
+			"A powerful data fetching and caching tool. It is designed to simplify common cases for loading data in a web application, eliminating the need to hand-write data fetching & caching logic yourself.",
+		image: "https://github.com/reduxjs.png",
+		tags: ["data fetching", "redux"],
 	},
 	{
 		name: "MobX",
@@ -679,10 +701,30 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		gitHubRepo: "pmndrs/zustand",
 		npmPackage: "zustand",
 		href: "https://docs.pmnd.rs/zustand/introduction",
-		image: "https://github.com/pmndrs.png",
+		image: "https://docs.pmnd.rs/zustand-resources/bear.png",
 		tags: ["state management"],
 		description:
 			"A small, fast and scalable bearbones state-management solution using simplified flux principles.",
+	},
+	{
+		name: "Jotai",
+		author: "Poimandres",
+		gitHubRepo: "pmndrs/jotai",
+		npmPackage: "jotai",
+		href: "https://jotai.org/",
+		image: "https://storage.googleapis.com/candycode/jotai/jotai-mascot.png",
+		tags: ["state management"],
+		description: "Primitive and flexible state management for React",
+	},
+	{
+		name: "Valtio",
+		author: "Poimandres",
+		gitHubRepo: "pmndrs/valtio",
+		npmPackage: "valtio",
+		href: "https://github.com/pmndrs/valtio",
+		image: "https://github.com/pmndrs.png",
+		tags: ["state management"],
+		description: "Lightweight Proxy-Based State Management for ReactJS and VanillaJS",
 	},
 	{
 		name: "A11y",
@@ -750,7 +792,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		description:
 			"A graphical user interface for changing variable states in React.",
 	},
-        {
+	{
 		name: "Blitz",
 		author: "Brandon Bayer",
 		gitHubRepo: "blitz-js/blitz",
@@ -774,7 +816,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 			"https://github.com/remix-run.png",
 		tags: ["framework", "routing", "data fetching", "server side rendering", "forms"],
 	},
-  	{
+	{
 		name: "Bumbag",
 		author: "Jake Moxey",
 		gitHubRepo: "jxom/bumbag-ui",
@@ -784,7 +826,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		tags: ["component library"],
 		description: "Build themeable React & React Native applications with your Bumbag",
 	},
-        {
+	{
 		name: "Chakra UI",
 		author: "Segun Adebayo",
 		gitHubRepo: "chakra-ui/chakra-ui",
@@ -794,7 +836,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		tags: ["component library"],
 		description: "⚡️ Simple, Modular & Accessible UI Components for your React Applications",
 	},
-        {
+	{
 		name: "Reach UI",
 		author: "Reach",
 		gitHubRepo: "reach/reach-ui",
@@ -804,7 +846,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		tags: ["component library"],
 		description: "The Accessible Foundation for React Apps and Design Systems",
 	},
-        {
+	{
 		name: "HashiCorp React Component",
 		author: "HashiCorp",
 		gitHubRepo: "hashicorp/react-components",
@@ -814,7 +856,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		tags: ["component library"],
 		description: "A set of HashiCorp branded react components currently maintained and used by the marketing dev team",
 	},
-        {
+	{
 		name: "React Spectrum",
 		author: "Adobe, Inc.",
 		gitHubRepo: "adobe/react-spectrum",
@@ -824,7 +866,7 @@ export const libraries: Library<typeof libraryTags[number]>[] = [
 		tags: ["component library"],
 		description: "A collection of libraries and tools that help you build adaptive, accessible, and robust user experiences.",
 	},
-        {
+	{
 		name: "Windmill React UI",
 		author: "Estevan Maito",
 		gitHubRepo: "estevanmaito/windmill-react-ui",