+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
+- [2021-03-14] Add Redux Toolkit, RTK Query, Jotai, Valtio
+  [\#169](https://github.com/thisdot/framework.dev/pull/169)
+  ([phryneas](https://github.com/phryneas))
 - [2022-03-13] Add Bumbag, Chakra UI, React Spectrum, Hashicorp React Component, Windmill React UI, Reach UI for React
   [\#175](https://github.com/thisdot/framework.dev/pull/175)
   ([sprabowo](https://github.com/sprabowo))
+- [2021-03-11] Add drop shadow to feature highlights
+  [\#174](https://github.com/thisdot/framework.dev/pull/174)
+- [2021-03-11] Fix iOS Safari dispaly bugs
+  [\#173](https://github.com/thisdot/framework.dev/pull/173)
+  ([jbachhardie](https://github.com/jbachhardie))
 - [2022-03-10] Fix the link "View all tools" on the homepage
   [\#171](https://github.com/thisdot/framework.dev/pull/171)
   ([RinatValiullov](https://github.com/RinatValiullov))